+++ NEWS.md
@@ -1,4 +1,5 @@
 # Changes in worldmet 0.6 [2016-01-15]
 
 - Add ability to return precipitation measurements, if available.
+- Add precipitation to `exportADMS`